feat: add support for unique GA events
Providing `once: true` to GoogleAnalytics::Action now calculates a
signature for the event based on its name and data. The `gtag_action`
plugin JS function checks to see if the event has been sent before
(based on its signature) and, if so, does not subsequently send the
event.

When `once` is `false` (the default, for backward compatibility), events
will be unconditionally sent.